moved menu and removed explore-container
