--- src/app/explore-container/explore-container.component.ts
@@ -1,11 +1,0 @@
-import {Component, Input} from "@angular/core";
-
-@Component({
-  selector: "app-explore-container",
-  templateUrl: "./explore-container.component.html",
-  styleUrls: ["./explore-container.component.scss"],
-  standalone: true,
-})
-export class ExploreContainerComponent {
-  @Input() name?: string;
-}
