+++ src/app/shared/components/menu/menu.component.spec.ts
@@ -1,7 +1,7 @@
 import {ComponentFixture, TestBed} from "@angular/core/testing";
 import {MenuComponent} from "./menu.component";
 import {RouterTestingModule} from "@angular/router/testing";
-import {AuthService} from "../core/services/auth.service";
+import {AuthService} from "../../../core/services/auth.service";
 import {of} from "rxjs";
 
 describe("MenuComponent", () => {
